move scss from layout to sass-components
