+++ project-new/src/components/Card-components/Photo.js
@@ -1,5 +1,6 @@
 import React from "react";
-import "../../stylesheets/Layout/Photo.scss";
+import "../../stylesheets/Sass-components/Photo.scss";
+
 
 function Photo() {
   return <div className="js__profile-image card__img-container"></div>;
