+++ project-new/src/components/Footer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import logoFooter from "../images/logo-adalab-80px.png";
-import '../stylesheets/Layout/Footer.scss';
+import '../stylesheets/Sass-components/Footer.scss';
 
 function Footer() {
     return (
