+++ project-new/src/stylesheets/Sass-components/Photo.scss
