+++ project-new/src/stylesheets/Sass-components/Reset.scss
@@ -1,4 +1,5 @@
 @import "../Core/variables.scss";
+
 .card__reset {
   margin-bottom: 2rem;
   border-radius: 1.5rem;
